# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Continuous LLM Evaluation & Regression Testing Platform API",
    "version": "0.1.0-draft",
    "summary": "Release-gate system of record for AI quality.",
    "description": "Specification only; no implementation exists. Serialized as JSON rather than YAML so the contract is validatable with the Python standard library, which keeps Phase 3 free of a dependency manifest.\n\nEvery operation and schema carries an `x-requirements` extension naming the requirements it realises. Traceability is therefore derived from this document mechanically rather than restated in a matrix that can drift.",
    "license": {
      "name": "Proprietary",
      "identifier": "LicenseRef-Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://{host}/v1",
      "description": "Versioned API root",
      "variables": {
        "host": {
          "default": "localhost:8000"
        }
      }
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/projects/{projectId}/runs": {
      "parameters": [
        {
          "$ref": "#/components/parameters/projectId"
        }
      ],
      "post": {
        "operationId": "createRun",
        "summary": "Submit an evaluation run",
        "description": "Freezes run identity before execution begins, estimates cost, and reserves budget. Rejects rather than partially executes when the estimate exceeds budget.",
        "tags": [
          "runs"
        ],
        "x-requirements": [
          "REQ-F-07-1",
          "REQ-F-10-5",
          "REQ-X-9",
          "REQ-N-REL-2",
          "REQ-N-SEC-1"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/idempotencyKey"
          },
          {
            "$ref": "#/components/parameters/correlationId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Run accepted; identity frozen.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "409": {
            "description": "Rejected before execution because the cost estimate exceeded the budget.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "503": {
            "$ref": "#/components/responses/PlatformFailure"
          }
        }
      }
    },
    "/runs/{runId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/runId"
        }
      ],
      "get": {
        "operationId": "getRun",
        "summary": "Retrieve a run with its completeness and reproducibility state",
        "tags": [
          "runs"
        ],
        "x-requirements": [
          "REQ-X-1",
          "REQ-N-COMP-2",
          "REQ-F-07-6"
        ],
        "responses": {
          "200": {
            "description": "The run.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/runs/{runId}/cancel": {
      "parameters": [
        {
          "$ref": "#/components/parameters/runId"
        }
      ],
      "post": {
        "operationId": "cancelRun",
        "summary": "Cancel an in-flight run",
        "description": "Leaves a consistent and clearly incomplete record.",
        "tags": [
          "runs"
        ],
        "x-requirements": [
          "REQ-F-07-7",
          "REQ-X-1"
        ],
        "responses": {
          "200": {
            "description": "Cancellation accepted; the run settles to the cancelled terminal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "The run already reached a terminal state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{runId}/samples": {
      "parameters": [
        {
          "$ref": "#/components/parameters/runId"
        }
      ],
      "get": {
        "operationId": "listRunSamples",
        "summary": "List per-sample outcomes",
        "description": "A sample that failed, timed out, or abstained carries a resolution state and no numeric score.",
        "tags": [
          "runs"
        ],
        "x-requirements": [
          "REQ-F-07-2",
          "REQ-X-2",
          "REQ-X-8",
          "REQ-F-01-5"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/cursor"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of samples.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSamplePage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/projects/{projectId}/gate-evaluations": {
      "parameters": [
        {
          "$ref": "#/components/parameters/projectId"
        }
      ],
      "post": {
        "operationId": "evaluateGate",
        "summary": "Evaluate a candidate against an approved baseline and return a gate decision",
        "description": "The CI entry point. Resolves the approved baseline without requiring the caller to restate it. Returns a decision whose outcome distinguishes platform failure from every quality outcome, and which always carries the evidence behind it.",
        "tags": [
          "gates"
        ],
        "x-requirements": [
          "REQ-F-09-1",
          "REQ-F-09-2",
          "REQ-F-09-4",
          "REQ-F-09-5",
          "REQ-F-09-7",
          "REQ-F-09-8",
          "REQ-F-08-7",
          "REQ-X-10",
          "REQ-F-01-3"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/idempotencyKey"
          },
          {
            "$ref": "#/components/parameters/correlationId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GateEvaluationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A gate decision was reached. Note that a decision of `hard_fail`, `insufficient_evidence`, or `not_comparable` is a successful API call: the platform did its job. Only `503` indicates the platform itself failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GateDecision"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "503": {
            "$ref": "#/components/responses/PlatformFailure"
          }
        }
      }
    },
    "/gate-decisions/{gateDecisionId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/gateDecisionId"
        }
      ],
      "get": {
        "operationId": "getGateDecision",
        "summary": "Retrieve a past gate decision and its evidence",
        "description": "Answers, for any past release decision, what evidence supported it, who approved it, and under which policy and statistical method version.",
        "tags": [
          "gates"
        ],
        "x-requirements": [
          "REQ-N-COMP-1",
          "REQ-F-09-4",
          "REQ-F-09-8"
        ],
        "responses": {
          "200": {
            "description": "The decision.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GateDecision"
                }
              },
              "text/markdown": {
                "schema": {
                  "type": "string",
                  "description": "The same decision as prose, carrying the same evidence: every criterion, the rule that fired, the interval and the sample size."
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/projects/{projectId}/baselines": {
      "parameters": [
        {
          "$ref": "#/components/parameters/projectId"
        }
      ],
      "post": {
        "operationId": "createBaseline",
        "summary": "Designate an approved baseline",
        "description": "Pins dataset, prompt, model configuration, suite, evaluator and judge versions. A change to any pinned version makes a later comparison not comparable.",
        "tags": [
          "baselines"
        ],
        "x-requirements": [
          "REQ-F-01-2",
          "REQ-X-4",
          "REQ-F-06-5"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaselineRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Baseline created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Baseline"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Scoped credential. Every authorization decision is enforced server-side."
      }
    },
    "parameters": {
      "projectId": {
        "name": "projectId",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/Ulid"
        }
      },
      "runId": {
        "name": "runId",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/Ulid"
        }
      },
...
      "idempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Duplicate delivery must not double-count results or cost.",
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 255
        },
        "x-requirements": [
          "REQ-N-REL-2"
        ]
      },
...
    },
...
  }
}